Updated evaluator to use a dict for math exp
Figures that there's a better way. Operators module:

https://docs.python.org/2/library/operator.html

I mapped the string * to the method operator.mul (etc). It's important to note that {' * ': operator.mul() } ~will not~ work because python is trying to do the function call inside the dictionary (which is weird, except not); however, referencing the operator.mul you can later pass it values to evaluate.